rejigged tests so that coverage worked
