+++ internal/test/support/shared_db.go
@@ -5,6 +5,7 @@ import (
 	"database/sql"
 	"fmt"
 	"log"
+	"os"
 	"strings"
 	"sync"
 	"time"
@@ -111,8 +112,18 @@ func SetupTestDBShared() *store.TokenisationStore {
 
 	log.Printf("Using shared database: %s\n", newConnStr)
 
+	paths := []string{"../../../db/migrations", "../../db/migrations", "../../../../db/migrations"}
+
+	var validPath string
+
+	for _, p := range paths {
+		if folderExists(p) {
+			validPath = p
+		}
+	}
+
 	tokenisationStore, err := store.NewTokenisationStore(newConnStr, config.Config{
-		MigrationsPath: "../../../../db/migrations",
+		MigrationsPath: validPath,
 	})
 	if err != nil {
 		log.Fatalf("Failed to create tokenisation store: %v", err)
@@ -124,4 +135,12 @@ func SetupTestDBShared() *store.TokenisationStore {
 	}
 
 	return tokenisationStore
+}
+
+func folderExists(path string) bool {
+	info, err := os.Stat(path)
+	if os.IsNotExist(err) {
+		return false
+	}
+	return err == nil && info.IsDir()
 }
