+++ pkg/store/invoices_test.go
@@ -1,4 +1,4 @@
-package store
+package store_test
 
 import (
 	"testing"
