+++ pkg/store/payments_test.go
@@ -1,4 +1,4 @@
-package store
+package store_test
 
 import (
 	"database/sql"
